requirements.txt
tweepy==4.14.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
nltk>=3.8.1
torch>=2.2.0
transformers>=4.37.0
seaborn>=0.13.0
matplotlib>=3.8.0
flask>=3.0.0
streamlit>=1.31.0
python-dotenv>=1.0.0
